1. Removed Main class in roll pair and using bootstrap. Relative modifications included. 2. Moved scala roll pair files from component to module root
