+++ bin/roll_pair/egg_pair_bootstrap.sh
@@ -96,7 +96,7 @@ else
 fi
 
 export PYTHONPATH=$PYTHONPATH:${pythonpath}
-cmd="${PYTHON} ${filepath} --session-id ${session_id} --server-node-id ${server_node_id} --cluster-manager ${cluster_manager_host}:${cluster_manager_port} --node-manager ${node_manager_port} --processor-id ${processor_id}"
+cmd="${PYTHON} ${filepath} --config ${config} --session-id ${session_id} --server-node-id ${server_node_id} --cluster-manager ${cluster_manager_host}:${cluster_manager_port} --node-manager ${node_manager_port} --processor-id ${processor_id}"
 
 if [[ -n ${port} ]]; then
   cmd="${cmd} --port ${port}"
