+++ conf/eggroll.properties
@@ -38,5 +38,5 @@ eggroll.rollpair.bootstrap.egg_pair.filepath=python/eggroll/roll_pair/egg_pair.p
 eggroll.rollpair.bootstrap.roll_pair_master.exepath=bin/roll_pair/roll_pair_master_bootstrap.sh
 eggroll.rollpair.bootstrap.roll_pair_master.javahome=
 eggroll.rollpair.bootstrap.roll_pair_master.classpath=jvm/core/target/classes:jvm/core/target/lib/*:jvm/roll_pair/target/classes:jvm/roll_pair/target/lib/*
-eggroll.rollpair.bootstrap.roll_pair_master.mainclass=com.webank.eggroll.rollpair.Main
+eggroll.rollpair.bootstrap.roll_pair_master.mainclass=com.webank.eggroll.rollpair.RollPairMasterBootstrap
 eggroll.rollpair.bootstrap.roll_pair_master.jvm.options=
