+++ jvm/roll_pair/main/scala/com/webank/eggroll/rollpair/client/RollPair.scala
@@ -28,7 +28,7 @@ import com.webank.eggroll.core.datastructure.{Broker, LinkedBlockingBroker}
 import com.webank.eggroll.core.meta.{ErJob, ErStore, ErStoreLocator}
 import com.webank.eggroll.core.session.{ErConf, RuntimeErConf}
 import com.webank.eggroll.core.transfer.GrpcTransferClient
-import com.webank.eggroll.rollpair.component.RollPairMaster
+import com.webank.eggroll.rollpair.RollPairMaster
 
 class RollPair(val store: ErStore, val opts: ErConf = RuntimeErConf()) {
   private var __store: ErStore = null
